refactor(src-stores): Rename ETHProvider to Provider
