+++ src/stores/Provider.ts
@@ -12,6 +12,7 @@ import { schema } from '../services/ABIService';
 
 export enum ContractType {
     ERC20 = 'ERC20',
+    DecentralizedAutonomousTrust = 'ERC20',
     Avatar = 'Avatar',
     Controller = 'Controller',
     Reputation = 'Reputation',
